ai-server/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
haystack-ai==2.6.1
transformers[torch,sentencepiece]==4.46.1
